Refactoring package and class names.
