--- app/src/main/java/in/developingdeveloper/timeline/add/event/domain/repositories/AddEventRepository.kt
@@ -1,7 +1,0 @@
-package `in`.developingdeveloper.timeline.add.event.domain.repositories
-
-import `in`.developingdeveloper.timeline.core.domain.event.models.Event
-
-interface AddEventRepository {
-    suspend fun addEvent(event: Event)
-}
