--- app/src/main/java/in/developingdeveloper/timeline/add/event/domain/usecases/AddEventUseCase.kt
@@ -1,7 +1,0 @@
-package `in`.developingdeveloper.timeline.add.event.domain.usecases
-
-import `in`.developingdeveloper.timeline.core.domain.event.models.Event
-
-interface AddEventUseCase {
-    suspend operator fun invoke(event: Event): Result<Unit>
-}
