--- app/src/main/java/in/developingdeveloper/timeline/add/event/ui/components/tags/bottomsheet/TagListBottomSheetDefaults.kt
@@ -1,5 +1,0 @@
-package `in`.developingdeveloper.timeline.add.event.ui.components.tags.bottomsheet
-
-object TagListBottomSheetDefaults {
-    const val maxHeightPercentage = 0.7f
-}
