--- app/src/main/java/in/developingdeveloper/timeline/add/event/ui/models/SelectableUITag.kt
@@ -1,7 +1,0 @@
-package `in`.developingdeveloper.timeline.add.event.ui.models
-
-data class SelectableUITag(
-    val id: String,
-    val label: String,
-    val isSelected: Boolean,
-)
