--- app/src/main/java/in/developingdeveloper/timeline/add/tag/domain/repositories/DefaultAddTagRepository.kt
@@ -1,13 +1,0 @@
-package `in`.developingdeveloper.timeline.add.tag.domain.repositories
-
-import `in`.developingdeveloper.timeline.core.data.local.tags.TagDataSource
-import `in`.developingdeveloper.timeline.core.domain.tags.models.Tag
-import javax.inject.Inject
-
-class DefaultAddTagRepository @Inject constructor(
-    private val tagDataSource: TagDataSource,
-) : AddTagRepository {
-    override suspend fun addTag(tag: Tag) {
-        tagDataSource.insertTag(tag)
-    }
-}
