--- app/src/main/java/in/developingdeveloper/timeline/add/tag/ui/models/AddTagViewState.kt
@@ -1,19 +1,0 @@
-package `in`.developingdeveloper.timeline.add.tag.ui.models
-
-data class AddTagViewState(
-    val form: NewTagForm,
-    val isLoading: Boolean,
-    val isFormEnabled: Boolean,
-    val isCompleted: Boolean,
-    val errorMessage: String?,
-) {
-    companion object {
-        val Initial = AddTagViewState(
-            form = NewTagForm.Initial,
-            isLoading = false,
-            isFormEnabled = true,
-            isCompleted = false,
-            errorMessage = null,
-        )
-    }
-}
