--- app/src/main/java/in/developingdeveloper/timeline/add/tag/ui/models/NewTagForm.kt
@@ -1,9 +1,0 @@
-package `in`.developingdeveloper.timeline.add.tag.ui.models
-
-data class NewTagForm(
-    val label: String,
-) {
-    companion object {
-        val Initial = NewTagForm(label = "")
-    }
-}
