+++ app/src/main/java/in/developingdeveloper/timeline/core/di/TagsModule.kt
@@ -4,14 +4,14 @@ import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
-import `in`.developingdeveloper.timeline.add.tag.domain.repositories.AddTagRepository
-import `in`.developingdeveloper.timeline.add.tag.domain.repositories.DefaultAddTagRepository
-import `in`.developingdeveloper.timeline.add.tag.domain.usecases.AddTagUseCase
-import `in`.developingdeveloper.timeline.add.tag.domain.usecases.DefaultAddTagUseCase
 import `in`.developingdeveloper.timeline.core.data.local.tags.RoomTagDataSource
 import `in`.developingdeveloper.timeline.core.data.local.tags.TagDataSource
 import `in`.developingdeveloper.timeline.core.domain.tags.repositories.DefaultTagRepository
 import `in`.developingdeveloper.timeline.core.domain.tags.repositories.TagRepository
+import `in`.developingdeveloper.timeline.modify.tag.domain.repositories.AddTagRepository
+import `in`.developingdeveloper.timeline.modify.tag.domain.repositories.DefaultAddTagRepository
+import `in`.developingdeveloper.timeline.modify.tag.domain.usecases.AddTagUseCase
+import `in`.developingdeveloper.timeline.modify.tag.domain.usecases.DefaultAddTagUseCase
 import `in`.developingdeveloper.timeline.taglist.domain.usecases.DefaultGetAllTagsUseCase
 import `in`.developingdeveloper.timeline.taglist.domain.usecases.GetAllTagsUseCase
 import javax.inject.Singleton
