+++ app/src/main/java/in/developingdeveloper/timeline/modify/event/domain/repositories/AddEventRepository.kt
@@ -1,0 +1,7 @@
+package `in`.developingdeveloper.timeline.modify.event.domain.repositories
+
+import `in`.developingdeveloper.timeline.core.domain.event.models.Event
+
+interface AddEventRepository {
+    suspend fun addEvent(event: Event)
+}
