+++ app/src/main/java/in/developingdeveloper/timeline/modify/event/domain/usecases/AddEventUseCase.kt
@@ -1,0 +1,7 @@
+package `in`.developingdeveloper.timeline.modify.event.domain.usecases
+
+import `in`.developingdeveloper.timeline.core.domain.event.models.Event
+
+interface AddEventUseCase {
+    suspend operator fun invoke(event: Event): Result<Unit>
+}
