+++ app/src/main/java/in/developingdeveloper/timeline/modify/event/ui/components/tags/bottomsheet/TagListBottomSheetDefaults.kt
@@ -1,0 +1,5 @@
+package `in`.developingdeveloper.timeline.modify.event.ui.components.tags.bottomsheet
+
+object TagListBottomSheetDefaults {
+    const val maxHeightPercentage = 0.7f
+}
