+++ app/src/main/java/in/developingdeveloper/timeline/modify/event/ui/models/ModifyEventViewState.kt
@@ -1,0 +1,23 @@
+package `in`.developingdeveloper.timeline.modify.event.ui.models
+
+data class ModifyEventViewState(
+    val form: ModifyEventForm,
+    val tagListViewState: SelectableTagListViewState,
+    val modifyTags: Boolean,
+    val formEnabled: Boolean,
+    val isLoading: Boolean,
+    val isCompleted: Boolean,
+    val errorMessage: String?,
+) {
+    companion object {
+        val Initial = ModifyEventViewState(
+            form = ModifyEventForm.Initial,
+            tagListViewState = SelectableTagListViewState.Initial,
+            modifyTags = false,
+            formEnabled = true,
+            isLoading = false,
+            isCompleted = false,
+            errorMessage = null,
+        )
+    }
+}
