+++ app/src/main/java/in/developingdeveloper/timeline/modify/event/ui/models/SelectableUITag.kt
@@ -1,0 +1,7 @@
+package `in`.developingdeveloper.timeline.modify.event.ui.models
+
+data class SelectableUITag(
+    val id: String,
+    val label: String,
+    val isSelected: Boolean,
+)
