+++ app/src/main/java/in/developingdeveloper/timeline/modify/tag/domain/repositories/AddTagRepository.kt
@@ -1,0 +1,7 @@
+package `in`.developingdeveloper.timeline.modify.tag.domain.repositories
+
+import `in`.developingdeveloper.timeline.core.domain.tags.models.Tag
+
+interface AddTagRepository {
+    suspend fun addTag(tag: Tag)
+}
