+++ app/src/main/java/in/developingdeveloper/timeline/modify/tag/domain/usecases/AddTagUseCase.kt
@@ -1,0 +1,7 @@
+package `in`.developingdeveloper.timeline.modify.tag.domain.usecases
+
+import `in`.developingdeveloper.timeline.core.domain.tags.models.Tag
+
+interface AddTagUseCase {
+    suspend operator fun invoke(tag: Tag): Result<Unit>
+}
