+++ app/src/main/java/in/developingdeveloper/timeline/modify/tag/ui/models/NewTagForm.kt
@@ -1,0 +1,9 @@
+package `in`.developingdeveloper.timeline.modify.tag.ui.models
+
+data class NewTagForm(
+    val label: String,
+) {
+    companion object {
+        val Initial = NewTagForm(label = "")
+    }
+}
